lab5/budget_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-jose==3.3.0
pymongo==3.12.0
